refactor: rename button component file and update import path
