+++ frontend/components/ui/Button.example.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import Button from './Button';
+import Button from './buttons';
 
 /**
  * Example usage of the Button component
